connector: fix a bug connection without communication
when a connection is established, it will setup a
timer to wait data, if timer timeouts before the data coming,
the timer will treat this connection as vicious and will clear it.
if the connection is closed before the timer exceeding, there is a
bug that the server does not clear the timer, this commit fixs it.

Change-Id: I025d29d72ffe00a84a83e8b717e58ebc083b51c1